requirements.txt
certifi==2017.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.0.11
Django==1.11.7
django-bootstrap-pagination==1.6.3
django-bootstrap3==9.1.0
django-bootstrap4==0.0.4
django-cacheops==4.0.1
django-filter==1.1.0
django-filters==0.2.1
django-jquery==3.1.0
django-oauth-toolkit==1.0.0
django-rest-swagger==2.1.2
django-widget-tweaks==1.4.1
djangorestframework==3.7.1
djangorestframework-jwt==1.11.0
ez-setup==0.9
ffmpy==0.2.2
funcy==1.9.1
idna==2.6
imageio==2.1.2
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
moviepy==0.2.3.2
Naked==0.1.31
numpy==1.13.3
oauthlib==2.0.6
olefile==0.44
openapi-codec==1.3.2
Pillow==4.3.0
pycrypto==2.6.1
PyJWT==1.5.3
pytz==2017.3
PyYAML==3.12
redis==2.10.6
requests==2.11.1
requests-toolbelt==0.7.0
shellescape==3.4.1
simplejson==3.12.0
six==1.11.0
tqdm==4.11.2
uritemplate==3.0.0
urllib3==1.22
